split DbScopeFactory into a generic and non-generic types
